refactor: restructure project into feature-based architecture
