+++ ai-learning-platform/apps/web/features/chat/types.ts
@@ -1,0 +1,12 @@
+export type ChatRole = "user" | "assistant" | "system";
+
+export type ChatMessage = {
+  id: string;
+  role: "user" | "assistant";
+  content: string;
+};
+
+export type IncomingChatMessage = {
+  role: ChatRole;
+  content: string;
+};
