+++ ai-learning-platform/apps/web/lib/openai/client.ts
@@ -1,0 +1,7 @@
+import OpenAI from "openai";
+
+export const openAiApiKey = process.env.OPENAI_API_KEY;
+
+export const openai = new OpenAI({
+  apiKey: openAiApiKey,
+});
